separate validation check logic
